fix plotting format and style tweaks
