--- data/histograms.py
@@ -1,22 +1,0 @@
-import matplotlib.pyplot as plt
-import pandas as pd
-import numpy as np
-
-# Generate data on commute times.
-size, scale = 1000, 10
-d = pd.Series(np.random.gamma(scale, size=size) ** 1.5)
-
-# An "interface" to matplotlib.axes.Axes.hist() method
-n, bins, patches = plt.hist(x=d, bins='auto', color='#0504aa',
-                            alpha=0.7, rwidth=0.85)
-
-plt.grid(axis='y', alpha=0.75)
-plt.xlabel('Value')
-plt.ylabel('Frequency')
-plt.title('My Very Own Histogram')
-plt.text(23, 45, r'$\mu=15, b=3$')
-maxfreq = n.max()
-# Set a clean upper y-axis limit.
-plt.ylim(ymax=np.ceil(maxfreq / 10) * 10 if maxfreq % 10 else maxfreq + 10)
-
-plt.show()
