+++ templates/plot_counts.html
@@ -20,7 +20,9 @@
           "{{ row['outcome_color']}}",
           "{{ row['outcome_datetime']}}",
           "{{ row['los']}}",
-          "{{ row['days_old']}}"
+          "{{ row['days_old']}}",
+          "{{ row['found_location']}}",
+          "{{ row['age_upon_outcome']}}"
           ]);
     {% endfor %}
 
@@ -37,15 +39,15 @@
 </script>
 
 <div class="col-lg-3">
-  <h3> choose custom filter: </h3>
+  <h3> Choose a custom filter: </h3>
   <div class="input-group mb-2">
     <form action="{{ url_for('counts_plot_filter') }}">
-      <input type="text" class="form-control" placeholder="i.e. animal_type" name="field" aria-label="field" aria-describedby="basic-addon1"><br>
-      <input type="text" class="form-control" placeholder="i.e. CAT" name="value" aria-label="value" aria-describedby="basic-addon2"><br>
+      <input type="text" class="form-control" placeholder="i.e. intake_type" name="field" aria-label="field" aria-describedby="basic-addon1"><br>
+      <input type="text" class="form-control" placeholder="i.e. Owner Surrender" name="value" aria-label="value" aria-describedby="basic-addon2"><br>
       <button class="btn btn-info btn-sm" type="submit">Submit</button>
     </form>
   </div>
-  <h3>Select your settings: </h3>
+  <h3>Select from available fields: </h3>
   <form class="form-inline" name="color_button" method='POST'>
     <br>
     <div class="input-group">
@@ -67,7 +69,7 @@
 </div>
 
 <div class="col-lg-9">
-  <h1>counts</h1>
+  <h1>Animal Count by Data Dimension</h1>
 
   <p> Filter by Animal Type:
     <a class="btn btn-info btn-sm" href="{{ url_for('counts_plot_filter', field='species',value='Cat') }}">Cats</a>
@@ -78,15 +80,16 @@
 
   <p> Filter by Outcome Type:
     <a class="btn btn-success btn-sm" href="{{ url_for('counts_plot_filter', field='outcome_type',value='Adoption') }}">Adoption</a>
-    <a class="btn btn-success btn-sm" href="{{ url_for('counts_plot_filter', field='outcome_type',value='Return To Owner') }}">Return to Owner</a>
+    <a class="btn btn-success btn-sm" href="{{ url_for('counts_plot_filter', field='outcome_type',value='Return to Owner') }}">Return to Owner</a>
     <a class="btn btn-success btn-sm" href="{{ url_for('counts_plot_filter', field='outcome_type',value='Euthanasia') }}">Euthanasia</a>
     <a class="btn btn-success btn-sm" href="{{ url_for('counts_plot_filter', field='outcome_type',value='Transfer') }}">Transfer</a>
   </p>
   <hr>
   {{ plot|safe }}
 </div>
+
 <hr>
-<div class="col-lg-9"></div>
+<div class="col-lg-7"></div>
   <hr>
   <table id="sc-table" class="table table-hover table-responsive">
     <thead>
@@ -102,6 +105,8 @@
         <th><label title="outcome_datetime">Outcome Date</label></th>
         <th><label title="los">Length of Stay</label></th>
         <th><label title="days_old">Days Old</label></th>
+        <th><label title="found_location">Found Location</label></th>
+        <th><label title="age_upon_outcome">Age at Outcome</label></th>
       </tr>
     </thead>
   </table>
