+++ templates/plot_mpld3.html
@@ -105,7 +105,7 @@
     <a class="btn btn-success btn-sm" href="{{ url_for('linked_plot_filter', field='outcome_type',value='Adoption') }}">Adoption</a>
     <a class="btn btn-success btn-sm" href="{{ url_for('linked_plot_filter', field='outcome_type',value='Returned to Owner') }}">Return to Owner</a>
     <a class="btn btn-success btn-sm" href="{{ url_for('linked_plot_filter', field='outcome_type',value='Euthanasia') }}">Euthanasia</a>
-    <a class="btn btn-success btn-sm" href="{{ url_for('linked_plot_filter', field='outcome_type',value='Transfer') }}">Foster Volunteer</a>
+    <a class="btn btn-success btn-sm" href="{{ url_for('linked_plot_filter', field='outcome_type',value='Transfer') }}">Transfer</a>
   </p>
 
   <p> Filter by Intake Type:
